// demo/admin/src/store/helpers/keys/types.ts
export const enum KeyTypes {
  NEW = 'new',
  ORG = 'org',
  ORGANISATIONS = 'organisations',
  ORGANISATION = 'organisation',
  ORGANISATIONBUDGET = 'organisationBudget',
  ORGANISATIONCOUNTRYBUDGET = 'organisationCountryBudget',
  ORGANISATIONDOC = 'organisationDoc',
  ORGANISATIONEXPENDITURE = 'organisationExpenditure',
  ORGANISATIONRECIPIENTBUDGET = 'organisationRecipientBudget',
  ORGANISATIONREGIONBUDGET = 'organisationRegionBudget',
  ACTIVITIES = 'activities',
  ACTIVITY = 'activity',
  ACTIVITYADDITIONAL = 'activityAdditional',
  ACTIVITYDATE = 'activityDate',
  ACTIVITYPARTICIPATINGORG = 'activityParticipatingOrg',
  ACTIVITYSECTOR = 'activitySector',
  ACTIVITYBUDGET = 'activityBudget',
  ACTIVITYTERRITORY = 'activityTerritory',
  ACTIVITYTRANSACTION = 'activityTransactions',
  ACTIVITYRELATEDACTIVITY = 'activityRelatedActivities'
}

export interface Keys {
  key: string
  keyType: KeyTypes
}